chore: minor improvements

+++ packages/shared/src/lib/core/layer-2/helpers/getEvmTransactionValueFromAmount.ts
@@ -2,5 +2,5 @@ import { HEX_PREFIX } from '@core/utils'
 import { WEI_PER_GLOW } from '../constants'
 
 export function getEvmTransactionValueFromAmount(amount: bigint): string {
-    return HEX_PREFIX + BigInt(amount * WEI_PER_GLOW).toString(16)
+    return HEX_PREFIX + (amount * WEI_PER_GLOW).toString(16)
 }
